+++ NodeJS-APIs/src/constrollers/livroController.js
@@ -1,68 +1,97 @@
 import livro from "../models/Livro.js";
+import NaoEncontrado from "../erros/naoEncontrado.js";
 import { autor } from "../models/Autor.js";
 
 class LivroController {
-    static async listarLivros (req, res) {
+    static listarLivros = async (req, res, next) => {
         try {
             const listaLivros = await livro.find({});
-            res.status(200).json(listaLivros);
+            if (listaLivros !== null) {
+                res.status(200).json(listaLivros);
+            } else {
+                next(new NaoEncontrado("Livros não localizados no banco de dados."));
+            }
         } catch (erro) {
-            res.status(500).json({message: `Erro ao listar livros: ${erro.message}`});
+            next(erro);
         }
     }
 
-    static async listarLivroPorId (req, res) {
+    static listarLivroPorId = async (req, res, next) => {
         try {
             const id = req.params.id;
             const livroEncontrado = await livro.findById(id);
-            res.status(200).json(livroEncontrado);
+            if (livroEncontrado !== null) {
+                res.status(200).json(livroEncontrado);
+            } else {
+                next(new NaoEncontrado("Livro não localizado no banco de dados."));
+            }
         } catch (erro) {
-            res.status(500).json({message: `Erro ao buscar livro: ${erro.message}`});
+            next(erro);
         }
     }
 
-    static async cadastrarLivros (req, res) {
+    static cadastrarLivros = async (req, res, next) => {
         const novoLivro = req.body;
         try {
             const autorEncontrado = await autor.findById(novoLivro.autor);
-            const livroCompleto = {...novoLivro, autor: {...autorEncontrado._doc}}
-            const livroCriado = await livro.create(livroCompleto);
-            res.status(201).json({message: "Livro criado com sucesso!", livro: livroCriado});
+            if (autorEncontrado !== null) {
+                const livroCompleto = {...novoLivro, autor: {...autorEncontrado._doc}}
+                const livroCriado = await livro.create(livroCompleto);
+                res.status(201).json({message: "Livro criado com sucesso!", livro: livroCriado});
+            } else {
+                next(new NaoEncontrado("Autor não localizado no banco de dados."));
+            }
         } catch (erro) {
-            res.status(500).json({message: `Erro ao cadastrar novo livro: ${erro.message}`});
+            next(erro);
         }        
     }
 
-    static async atualizarLivro (req, res) {
+    static atualizarLivro = async (req, res, next) => {
         const livroBody = req.body;
         try {
             const id = req.params.id;
-            const autorEncontrado = await autor.findById(livroBody.autor);
-            const livroCompleto = {...livroBody, autor: {...autorEncontrado._doc}}
-            await livro.findByIdAndUpdate(id, livroCompleto);
-            res.status(200).json({message: "Livro atualizado com sucesso."});
+            if (id !== null) {
+                const autorEncontrado = await autor.findById(livroBody.autor);
+                if (autorEncontrado !== null) {
+                    const livroCompleto = {...livroBody, autor: {...autorEncontrado._doc}}
+                    await livro.findByIdAndUpdate(id, livroCompleto);
+                    res.status(200).json({message: "Livro atualizado com sucesso."});
+                } else {
+                    next(new NaoEncontrado("Autor não localizado no banco de dados."));
+                }
+            } else {
+                next(new NaoEncontrado("Livro não localizado no banco de dados."));
+            }
         } catch (erro) {
-            res.status(500).json({message: `Erro ao atualizar livro: ${erro.message}`});
+            next(erro);
         }
     }
 
-    static async deletarLivro (req, res) {
+    static deletarLivro = async (req, res, next) => {
         try {
             const id = req.params.id;
-            await livro.findByIdAndDelete(id);
-            res.status(200).json({message: "Livro deletado com sucesso."});
+            if (id !== null) {
+                await livro.findByIdAndDelete(id);
+                res.status(200).json({message: "Livro deletado com sucesso."});
+            } else {
+                next(new NaoEncontrado("Livro não localizado no banco de dados."));
+            }            
         } catch (erro) {
-            res.status(500).json({message: `Erro ao deletar livro: ${erro.message}`});
+            next(erro);
         }
     }
 
-    static async listarLivrosPorEditora (req, res) {
+    static listarLivrosPorEditora = async (req, res, next) => {
         const editora = req.query.editora;
         try {
             const livrosPorEditora = await livro.find({editora: editora});
-            res.status(200).json(livrosPorEditora);
+            if (livrosPorEditora !== null) {
+                res.status(200).json(livrosPorEditora);
+            } else {
+                next(new NaoEncontrado("Livros não localizados no banco de dados."));
+            }
         } catch (erro) {
-            res.status(500).json({message: `Erro ao listar livros por editora: ${erro.message}`});
+            next(erro);
         }
     }
 }
